+++ CHANGELOG.md
@@ -5,10 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+
+## [15.2.0][15.2.0] - 2016-09-02
 ### Added
 - [Feature] Add `colSpan` prop to `Table.TD` component. (#533)
 - [Feature] Add size class based on icon size prop. (#517)
-- [Feature] Add a `density` prop to `Table` with a `loose` option. (#536) 
+- [Feature] Add a `density` prop to `Table` with a `loose` option. (#536)
 
 ### Fixed
 - [Patch] Revert TabNav examples, the minimal examples throw errors because those props are required.
@@ -515,7 +517,7 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 ### Added
 - [Patch] Use semantic versioning in LEGO (#58).
 
-[Unreleased]: https://github.com/optimizely/oui/compare/v15.1.0...HEAD
+[Unreleased]: https://github.com/optimizely/oui/compare/v15.2.0...HEAD
 [0.0.2]: https://github.com/optimizely/oui/compare/v0.0.1...v0.0.2
 [0.0.3]: https://github.com/optimizely/oui/compare/v0.0.2...v0.0.3
 [1.0.0]: https://github.com/optimizely/oui/compare/v0.0.3...v1.0.0
@@ -562,3 +564,4 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 [14.1.0]: https://github.com/optimizely/oui/compare/v14.0.5...v14.1.0
 [15.0.0]: https://github.com/optimizely/oui/compare/v14.1.0...v15.0.0
 [15.1.0]: https://github.com/optimizely/oui/compare/v15.0.0...v15.1.0
+[15.2.0]: https://github.com/optimizely/oui/compare/v15.1.0...v15.2.0